+++ src/app/cli/src/explore/api_server.rs
@@ -60,6 +60,14 @@ impl APIServer {
             )
             .nest("/", kamu_adapter_http::data::root_router())
             .nest(
+                "/odata",
+                if multi_tenant_workspace {
+                    kamu_adapter_odata::router_multi_tenant()
+                } else {
+                    kamu_adapter_odata::router_single_tenant()
+                },
+            )
+            .nest(
                 if multi_tenant_workspace {
                     "/:account_name/:dataset_name"
                 } else {