+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Experimental support for [OData](https://www.odata.org/) protocol
 ### Fixed
-- Pulling datasets by account in multitenant workspace
+- Pulling datasets by account in multi-tenant workspace
 
 ## [0.165.0] - 2024-03-12
 ### Updated